chore: purgecss-test
